control/requirements.txt
pynput
pyzmq
pyperclip